% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_all_linters.R
\name{use_all_linters}
\alias{use_all_linters}
\title{Setup a lintr configuration file with all linters explicitly enabled}
\usage{
use_all_linters(..., dir = ".", overwrite = FALSE)
}
\arguments{
\item{...}{Not currently used.}

\item{dir}{Directory to create the \code{.lintr.R} file within.}

\item{overwrite}{Should an existing \code{.lintr.R} file be overwritten.}
}
\description{
Setup a lintr configuration file with all linters explicitly enabled
}
\examples{
dir <- tempdir()
f <- try(use_all_linters(dir = dir))
if (!inherits(f, "try-error")) {
    unlink(f)
}

}
